Some improvements

+++ src/main/java/Commands/External/YouTubeMP3Command.java
@@ -3,6 +3,7 @@ package Commands.External;
 import CommandListeners.CommandProperties;
 import CommandListeners.onRecievedListener;
 import CommandSupporters.Command;
+import Constants.Permission;
 import General.*;
 import General.Internet.Internet;
 import General.Mention.MentionFinder;
@@ -29,6 +30,8 @@ import java.util.Optional;
         trigger = "ytmp3",
         withLoadingBar = true,
         emoji = "\uD83C\uDFB5",
+        botPermissions = Permission.ATTACH_FILES_TO_TEXT_CHANNEL,
+        userPermissions = Permission.ATTACH_FILES_TO_TEXT_CHANNEL,
         thumbnail = "http://icons.iconarchive.com/icons/martz90/circle/128/youtube-icon.png",
         executable = false,
         aliases = {"youtubemp3", "yt"}
@@ -41,33 +44,36 @@ public class YouTubeMP3Command extends Command implements onRecievedListener {
 
     @Override
     public boolean onRecieved(MessageCreateEvent event, String followedString) throws Throwable {
-        Optional<String> videoIdOptional = YouTubeDownloader.getVideoID(followedString);
 
-        if (videoIdOptional.isPresent()) {
-            String loadingEmoji = Tools.getLoadingReaction(event.getServerTextChannel().get());
-            Message message = event.getChannel().sendMessage(EmbedFactory.getCommandEmbedStandard(this, getString("loading", loadingEmoji))).get();
+        if (!followedString.isEmpty()) {
+            Optional<String> videoIdOptional = YouTubeDownloader.getVideoID(followedString);
 
-            try {
-                File audioFile = YouTubeDownloader.downloadAudio(videoIdOptional.get());
-
-                if (audioFile == null) {
-                    message.edit(EmbedFactory.getCommandEmbedError(this, getString("toolong_desc"), getString("toolong_title"))).get();
+            if (videoIdOptional.isPresent()) {
+                String videoId = videoIdOptional.get();
+                if (videoId.equalsIgnoreCase("%toolong")) {
+                    event.getChannel().sendMessage(EmbedFactory.getCommandEmbedError(this, getString("toolong_desc"), getString("toolong_title"))).get();
                     return false;
                 }
 
-                event.getChannel().sendMessage(event.getMessage().getUserAuthor().get().getMentionTag(), EmbedFactory.getCommandEmbedStandard(this, getString("finished")), audioFile).get();
-                audioFile.deleteOnExit();
-            } catch (IOException e) {
-                //Ignore
-                message.edit(EmbedFactory.getCommandEmbedError(this, getString("empty_desc"), getString("empty_title"))).get();
-                return false;
+                String loadingEmoji = Tools.getLoadingReaction(event.getServerTextChannel().get());
+                Message message = event.getChannel().sendMessage(EmbedFactory.getCommandEmbedStandard(this, getString("loading", loadingEmoji))).get();
+
+                try {
+                    File audioFile = YouTubeDownloader.downloadAudio(videoId);
+                    event.getChannel().sendMessage(event.getMessage().getUserAuthor().get().getMentionTag(), EmbedFactory.getCommandEmbedStandard(this, getString("finished")), audioFile).get();
+                    return true;
+                } catch (Throwable e) {
+                    throw e;
+                } finally {
+                    message.delete().get();
+                }
             }
 
-            message.delete().get();
-            return true;
+            event.getChannel().sendMessage(EmbedFactory.getCommandEmbedError(this, getString("invalid", followedString))).get();
+            return false;
+        } else {
+            event.getChannel().sendMessage(EmbedFactory.getCommandEmbedError(this, getString("noargs"))).get();
+            return false;
         }
-
-        event.getChannel().sendMessage(EmbedFactory.getCommandEmbedError(this, getString("invalid", followedString))).get();
-        return false;
     }
 }
